refactor/feat(skills-section): added sliding animation to grid rows if motion is enabled by user and separated components within this section to facilitate future maintanence
