+++ src/pages/whoiam.tsx
@@ -1,15 +1,14 @@
-import { VStack, Wrap, WrapItem } from "@chakra-ui/react";
+import { VStack } from "@chakra-ui/react";
 
-import { TransitionSection } from "../components/TransitionSection";
-import { Section } from "../components/Section";
-import { FadeInWhenVisible } from "../components/FadeInWhenVisible";
-import { MiniCard } from "../components/MiniCard";
-import { skills } from "../utils/skills";
-import { Paragraph } from "../components/Paragraph";
 import { GetStaticProps } from "next";
 import { CustomButton } from "../components/CustomButton";
 import { ExternalLink } from "../components/ExternalLink";
+import { FadeInWhenVisible } from "../components/FadeInWhenVisible";
+import { Paragraph } from "../components/Paragraph";
+import { Section } from "../components/Section";
 import { SEO } from "../components/SEO";
+import { Skills } from "../components/Skills";
+import { TransitionSection } from "../components/TransitionSection";
 
 export default function WhoIAm() {
   return (
@@ -22,14 +21,23 @@ export default function WhoIAm() {
         <Section title="Who I Am">
           <VStack spacing="3rem" w="100%" alignItems="start">
             <Paragraph>
-              I’m Wilson. With over 5 years of professional experience, I am a Full Stack Developer with a strong background in Front End (UI/UX) and Geographic Information Systems (GIS). I've built scalable, data-driven solutions - ranging from small applications to enterprise platforms serving over a million users - that are intuitive, accessible, and high-performing. 
-          </Paragraph>
+              I’m Wilson. With over 5 years of professional experience, I am a
+              Full Stack Developer with a strong background in Front End (UI/UX)
+              and Geographic Information Systems (GIS). I've built scalable,
+              data-driven solutions - ranging from small applications to
+              enterprise platforms serving over a million users - that are
+              intuitive, accessible, and high-performing.
+            </Paragraph>
             <Paragraph>
-              Worked in all stages of software development, including planning, designing, building, testing, and deployment.
-          </Paragraph>
+              Worked in all stages of software development, including planning,
+              designing, building, testing, and deployment.
+            </Paragraph>
             <Paragraph>
-              Technology and sectors that I have developed solutions for: Healthcare, Services, Quality Assurance (Data Visualization Tools), Geographic Information Systems (GIS), Higher Education, Logistics, and Environmental Science;
-          </Paragraph>
+              Technology and sectors that I have developed solutions for:
+              Healthcare, Services, Quality Assurance (Data Visualization
+              Tools), Geographic Information Systems (GIS), Higher Education,
+              Logistics, and Environmental Science;
+            </Paragraph>
             <ExternalLink href="https://www.linkedin.com/in/wilsonfsouza/?locale=en_US">
               <CustomButton>See Resume</CustomButton>
             </ExternalLink>
@@ -37,27 +45,11 @@ export default function WhoIAm() {
         </Section>
       </FadeInWhenVisible>
 
-      <Section title="Skills">
-        <Wrap
-          w="100%"
-          spacing={{ base: "1.25rem", sm: "1.875rem", md: "2.125rem" }}
-          display="flex"
-          alignItems="center"
-          justifyContent="flex-start"
-        >
-          {skills.map(skill => (
-            <WrapItem key={skill.title}>
-              <FadeInWhenVisible>
-                <MiniCard icon={skill.icon} title={skill.title} />
-              </FadeInWhenVisible>
-            </WrapItem>
-          ))}
-        </Wrap>
-      </Section>
+      <Skills />
 
       <TransitionSection gradientDirection="upsidedown" />
     </main>
-  )
+  );
 }
 
 const ONE_DAY = 60 * 60 * 24;
@@ -65,8 +57,6 @@ const ONE_DAY = 60 * 60 * 24;
 export const getStaticProps: GetStaticProps = async () => {
   return {
     props: {},
-    revalidate: ONE_DAY
-  }
-}
-
-
+    revalidate: ONE_DAY,
+  };
+};
