--- src/utils/skills.tsx
@@ -1,120 +1,0 @@
-import {
-    SiTypescript,
-    SiJavascript,
-    SiReact,
-    SiNextDotJs,
-    SiNodeDotJs,
-    SiPython,
-    SiFigma,
-    SiGit,
-    SiSass,
-    SiMongodb,
-    SiPostgresql,
-    SiDocker,
-    SiJest,
-    SiTailwindcss,
-    SiTerraform,
-    SiRedux,
-    SiWebpack,
-    SiGithubactions,
-    SiAzuredevops,
-    SiAmazonaws,
-    SiGooglecloud,
-    SiMaterialUi,
-    SiMicrosoftsqlserver,
-} from 'react-icons/si';
-
-export const skills = [
-    {
-        icon: SiTypescript,
-        title: "TypeScript",
-    },
-    {
-        icon: SiJavascript,
-        title: "JavaScript",
-    },
-    {
-        icon: SiReact,
-        title: "React",
-    },
-    {
-        icon: SiNextDotJs,
-        title: "Next.js",
-    },
-    {
-        icon: SiNodeDotJs,
-        title: "Node.js",
-    },
-    {
-        icon: SiPython,
-        title: "Python",
-    },
-    {
-        icon: SiFigma,
-        title: "Figma",
-    },
-    {
-        icon: SiGit,
-        title: "Git",
-    },
-    {
-        icon: SiMongodb,
-        title: "MongoDB",
-    },
-    {
-        icon: SiPostgresql,
-        title: "PostgreSQL",
-    },
-    {
-        icon: SiSass,
-        title: "SASS",
-    },
-    {
-        icon: SiDocker,
-        title: "Docker",
-    },
-    {
-        icon: SiJest,
-        title: "Jest",
-    },
-    {
-        icon: SiTailwindcss,
-        title: "Tailwind CSS",
-    },
-    {
-        icon: SiTerraform,
-        title: "Terraform",
-    },
-    {
-        icon: SiRedux,
-        title: "Redux",
-    },
-    {
-        icon: SiWebpack,
-        title: "Webpack",
-    },
-    {
-        icon: SiGithubactions,
-        title: "CI/CD",
-    },
-    {
-        icon: SiAzuredevops,
-        title: "Azure",
-    },
-    {
-        icon: SiAmazonaws,
-        title: "AWS",
-    },
-    {
-        icon: SiGooglecloud,
-        title: "GCP",
-    },
-    {
-        icon: SiMaterialUi,
-        title: "MUI",
-    },
-    {
-        icon: SiMicrosoftsqlserver,
-        title: "MS SQL",
-    },
-]
